+++ binding/src/main/kotlin/ch/viseon/threejs/declarations/cameras/OrthographicCamera.kt
@@ -1,22 +1,22 @@
-//Generated date 22.11.2018
+//Generated date 23.11.2018
 @file:JsModule("three")
 package ch.viseon.threejs.declarations.cameras
 
 	/**
 	* Camera that uses [link:https://en.wikipedia.org/wiki/Orthographic_projection orthographic projection]. In this projection mode, an object's size in the rendered image stays constant regardless of its distance from the camera. This can be useful for rendering 2D scenes and UI elements, amongst other things.
 	*/
-open external class OrthographicCamera(left: Float = definedExternally, right: Float = definedExternally, top: Float = definedExternally, bottom: Float = definedExternally, near: Float = definedExternally, far: Float = definedExternally): ch.viseon.threejs.declarations.cameras.Camera{
+open external class OrthographicCamera(left: Double = definedExternally, right: Double = definedExternally, top: Double = definedExternally, bottom: Double = definedExternally, near: Double = definedExternally, far: Double = definedExternally): ch.viseon.threejs.declarations.cameras.Camera{
 
 	/**
 	* Camera frustum bottom plane.
 	*/
-	open var bottom: Float  = definedExternally
+	open var bottom: Double  = definedExternally
 
 
 	/**
 	* Camera frustum far plane. Default is **2000**. The valid range is between the current value of the [page:.near near] plane and infinity.
 	*/
-	open var far: Float  = definedExternally
+	open var far: Double  = definedExternally
 
 
 	/**
@@ -28,25 +28,25 @@ open external class OrthographicCamera(left: Float = definedExternally, right: F
 	/**
 	* Camera frustum left plane.
 	*/
-	open var left: Float  = definedExternally
+	open var left: Double  = definedExternally
 
 
 	/**
 	* Camera frustum near plane. Default is **0.1**. The valid range is between 0 and the current value of the [page:.far far] plane. Note that, unlike for the [page:PerspectiveCamera], **0** is a valid value for an OrthographicCamera's near plane.
 	*/
-	open var near: Float  = definedExternally
+	open var near: Double  = definedExternally
 
 
 	/**
 	* Camera frustum right plane.
 	*/
-	open var right: Float  = definedExternally
+	open var right: Double  = definedExternally
 
 
 	/**
 	* Camera frustum top plane.
 	*/
-	open var top: Float  = definedExternally
+	open var top: Double  = definedExternally
 
 
 	/**
@@ -58,13 +58,13 @@ open external class OrthographicCamera(left: Float = definedExternally, right: F
 	/**
 	* Gets or sets the zoom factor of the camera. Default is **1**.
 	*/
-	open var zoom: Float  = definedExternally
+	open var zoom: Double  = definedExternally
 
 
 	/**
 	* fullWidth — full width of multiview setup fullHeight — full height of multiview setup x — horizontal offset of subcamera y — vertical offset of subcamera width — width of subcamera height — height of subcamera Sets an offset in a larger [link:https://en.wikipedia.org/wiki/Viewing_frustum viewing frustum]. This is useful for multi-window or multi-monitor/multi-machine setups. For an example on how to use it see [page:PerspectiveCamera.setViewOffset PerspectiveCamera].
 	*/
-	open fun setViewOffset(fullWidth: Float = definedExternally, fullHeight: Float = definedExternally, x: Float = definedExternally, y: Float = definedExternally, width: Float = definedExternally, height: Float = definedExternally) : dynamic/*null*/
+	open fun setViewOffset(fullWidth: Double = definedExternally, fullHeight: Double = definedExternally, x: Double = definedExternally, y: Double = definedExternally, width: Double = definedExternally, height: Double = definedExternally) : dynamic/*null*/
 
 
 	/**
